requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
google-genai>=1.5.0
pydantic>=2.0.0
ollama>=0.6.0
